First attempt at fixing the VO sample.
So it turns out there is a lot of fragility around fonts in different versions of unity.
And a lot of fragility around symlinks which is how a lot of this was developed.
This removes Montserrat because the fragility around fonts meant having two fonts was causing issues, so now that sample is just Noto Sans.