fixed the unity loc sample?
I think this is now fixed.
The issue is some sort of addressables difference between 2022.3 and 6.
This to the best of my knowledge should not have ever been a problem, and yet somehow it was.
The tldr is do any asset linking on 2022.3 and it will work fine in 6 but not the other way around.